pyproject.toml
[project]
name = "testcal"
version = "0.1.0"
description = "This project fetches calendar from google calendar"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.0.0",
    "google-auth-oauthlib>=0.4.6",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
testcal = "testcal:main"

[tool.ruff]
line-length = 88
target-version = "py313"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = ["E501"]  # line length violations

[tool.ruff.isort]
known-first-party = ["testcal"]

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]